# [Test] add perf test script