it get result like below
```
Perf Test(SYCL)
test/PlaidML/CppEdsl.Add.mlir
0.076000 ms
test/PlaidML/CppEdsl.BigDot.mlir
3.258240 ms
2373.171875 ms
test/PlaidML/CppEdsl.BitAnd.mlir
0.076000 ms
test/PlaidML/CppEdsl.BitAndScalar.mlir
0.077440 ms
test/PlaidML/CppEdsl.BitLeft.mlir
0.093760 ms
test/PlaidML/CppEdsl.BitNot.mlir
0.076480 ms
0.080640 ms
test/PlaidML/CppEdsl.BitOr.mlir
0.077120 ms

```




Please review these guidelines to help with the review process:
- [ ] Have you provided a meaningful PR description?
- [ ] Have you added a test, a reproducer, or a reference to an issue with a reproducer?
- [ ] Have you tested your changes locally for CPU and GPU devices?
- [ ] Have you made sure that new changes do not introduce compiler warnings?
- [ ] If this PR is a work in progress, are you filing the PR as a draft?
- [ ] Have you organized your commits logically and ensured each can be built by itself?